+++ management/custom_domain_test.go
@@ -11,7 +11,7 @@ import (
 func TestCustomDomain(t *testing.T) {
 
 	c := &CustomDomain{
-		Domain:             auth0.Stringf("auth.%d.alexkappa.com", time.Now().UTC().Unix()),
+		Domain:             auth0.Stringf("%d.auth.uat.alexkappa.com", time.Now().UTC().Unix()),
 		Type:               auth0.String("auth0_managed_certs"),
 		VerificationMethod: auth0.String("txt"),
 	}