+++ meta.go
@@ -1,4 +1,5 @@
 package auth0
 
-// Version of this library.
-var Version = "5.0.0"
+// Version of this library. This value is generated automatically during the
+// release process; DO NOT EDIT.
+var Version = "latest"